requirements.txt
pandas
numpy
pytz
ib_insync
textual
rich
pyyaml
requests
matplotlib
pytest
